requirements.txt
# Core LLM
openai>=1.0.0
python-dotenv>=1.0.0

# LangChain for Agent Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Music Processing - Phase 1 (Hearing Tool)
basic-pitch>=0.3.0
librosa>=0.10.0
midiutil>=1.2.1
pretty_midi>=0.2.10
numpy>=1.24.0
scipy>=1.11.0

# Audio file handling
soundfile>=0.12.0
audioread>=3.0.0

# Audio Generation (Phase 3 - Future)
# audiocraft>=1.0.0
